Fix 500s on history/stream: INT4 vs i64 type mismatch, and FK violations on queue inserts
HistoryEvent.indexer_id was Option<i64> but the DB column is INTEGER
(INT4), causing sqlx decode failures on every poll. Changed to i32.
Same fix for QueueItem.download_client_id and indexer_id.

Also use NULL for download_client_id when the grab comes from an
embedded engine (negative synthetic ID) to avoid FK constraint
violations against the download_clients table.

+++ crates/stackarr-core/src/models/history.rs
@@ -28,7 +28,7 @@ pub struct HistoryEvent {
     pub languages: Option<serde_json::Value>,
     pub source_title: String,
     pub download_id: Option<String>,
-    pub indexer_id: Option<i64>,
+    pub indexer_id: Option<i32>,
     pub download_client: Option<String>,
     pub data: Option<serde_json::Value>,
     pub occurred_at: DateTime<Utc>,
